Backend/requirements.txt
Flask==2.0.1
google-cloud-speech==2.9.0
Flask-CORS==3.0.10
